Structure capitalize and reverse-string
